Carry resolved Windows sandbox policy through wrapper
